added Tracker class in space which tracks active/inactive stuff

+++ src/montecarlo.h
@@ -109,7 +109,7 @@ class TranslationalEntropy {
   private:
     Space &trial_spc;                              //!< Space after proposed MC move ("trial")
     Space &spc;                                    //!< Space before MC move ("default")
-    double bias(int trial_count, int count) const; //!< Bias due to change in atom/molecule numbers
+    double bias(size_t trial_count, size_t count) const; //!< Bias due to change in atom/molecule numbers
     double atomSwapEnergy(const Change::GroupChange&); //!< Contribution from atomic swap move
     double atomChangeEnergy(int molid);            //!< Contribution from size-change of atomic group
     double moleculeChangeEnergy(int molid);        //!< Contribution frin change in number of molecular groups
